chore(pr) Clean up from PR feedback. - Renamed `starunstaricon.tsx` to `star-unstar-icon.tsx`. - Renamed an internal field of the Remix navigation bar.
