+++ editor/src/components/canvas/star-unstar-icon.tsx
